Revert "[Quantization] Consolidate experts_int8 with fp8 online quantization (#38463)"
This reverts commit ad2b1277f99f1cee3d59d687aa1813267299f6a2.
