+++ vllm/model_executor/layers/quantization/__init__.py
@@ -40,7 +40,6 @@ QuantizationMethods = Literal[
     # shorthand for creating a more complicated online quant config object
     "fp8_per_tensor",
     "fp8_per_block",
-    "int8_per_channel_weight_only",
 ]
 QUANTIZATION_METHODS: list[str] = list(get_args(QuantizationMethods))
 
@@ -48,6 +47,7 @@ DEPRECATED_QUANTIZATION_METHODS = [
     "tpu_int8",
     "fbgemm_fp8",
     "fp_quant",
+    "experts_int8",
 ]
 
 # The customized quantization methods which will be added to this dict.
